requirements.txt
torch
torchvision
opencv-python
mediapipe
Pillow
numpy
